pyproject.toml
[project]
name = "eag-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.40.0",
    "mcp[cli]>=1.15.0",
    "pillow>=11.3.0",
    "pywinauto>=0.6.9",
    "yfinance>=0.2.40",
    "requests>=2.32.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "faiss-cpu>=1.12.0",
    "youtube-transcript-api>=1.2.3",
    "yt-dlp>=2025.10.22",
    "pyyaml>=6.0.3",
    "markitdown>=0.1.3",
    "trafilatura>=2.0.0",
    "pymupdf4llm>=0.0.27",
    "tqdm>=4.67.1",
    "telethon>=1.41.2",
    "xdg-base-dirs>=6.0.2",
    "aiohttp>=3.13.2",
    "fastapi>=0.121.0",
    "google-auth-oauthlib>=1.2.3",
    "google-api-python-client>=2.187.0",
]
